ai-service/requirements.txt
# SmolVLM2 için özel transformers dalı (zorunlu)
git+https://github.com/huggingface/transformers@v4.49.0-SmolVLM-2

# PyTorch CUDA 12.1 desteği
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

# Görüntü işleme
Pillow==11.0.0
numpy==2.1.2

# FastAPI ve web servisi
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
uvloop==0.21.0
pydantic==2.11.7

# Database
asyncpg==0.30.0

# HTTP ve yardımcı
requests==2.32.5
tqdm==4.67.1

# Log ve monitoring
psutil==7.0.0

# Gerekli ek paketler
num2words==0.5.14

# Hugging Face paketleri
huggingface-hub==0.34.4
tokenizers==0.21.4
safetensors==0.6.2
accelerate>=0.26.0
